+++ activesupport/CHANGELOG.md
@@ -189,4 +189,8 @@
 
     *Daniel Schierbeck*
 
+*   DateTime `advance` now supports partial days (`days: 1.5`)
+
+    *Shay Davidson*
+
 Please check [4-0-stable](https://github.com/rails/rails/blob/4-0-stable/activesupport/CHANGELOG.md) for previous changes.